+++ server/src/main/scala/chess/domain/Coordinate.scala
@@ -1,9 +1,10 @@
 package com.chessonline
-package game.domain
+package chess.domain
 
-/** Represents a coordinate on a [[Chessboard]] that is located on an intersection
-  * of some row (called `rank`) and some column (called `file`). Used to identify
+/** Represents a coordinate on a [[Chessboard]] that is located at the intersection
+  * of some chessboard row (called `rank`) and some chessboard column (called `file`). Used to identify
   * [[Square]]s on a [[Chessboard]].
+  *
   * @param file a chessboard column (file) this coordinate belongs to.
   * @param rank a chessboard row (rank) this coordinate belongs to.
   */